AI_Score_Engine/requirements.txt
fastapi
uvicorn
numpy
requests
python-dotenv
pydantic
PyMuPDF
torch
transformers
sentence-transformers
scikit-learn
